backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.4.0
openpyxl==3.1.2
reportlab==4.0.4
pandas==2.2.2
numpy==2.1.3
django-extensions==3.2.3
django-debug-toolbar==4.2.0
pytest-django==4.5.2
factory-boy==3.3.0
requests==2.31.0
celery==5.3.1
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.5.0
python-json-logger==2.0.7
boto3==1.28.44
botocore==1.31.44
uvicorn==0.23.2
